requirements.txt
discord.py==2.3.2
aiohttp==3.8.6
python-dotenv==1.0.0
aiosqlite==0.19.0
audioop-lts==0.2.1
